# fix(#632) - Locale pipe optimization