--- a/libs/transloco-locale/src/lib/tests/pipes/transloco-decimal.pipe.spec.ts
+++ b/libs/transloco-locale/src/lib/tests/pipes/transloco-decimal.pipe.spec.ts
@@ -79,4 +79,31 @@ describe('TranslocoDecimalPipe', () => {
       expect(spectator.element).toHaveText('');
     });
   });
+
+  it('should return previous result with same config', () => {
+    spectator = pipeFactory(
+      `{{ data | translocoDecimal:config }}`,
+      {
+        hostProps: {
+          data: 123,
+          config: { useGrouping: false, maximumFractionDigits: 3 }
+        }
+      }
+    );
+
+    const [, { useGrouping, maximumFractionDigits }] = getIntlCallArgs();
+    expect(useGrouping).toBeFalsy();
+    expect(maximumFractionDigits).toEqual(3);
+    const first = spectator.element.textContent;
+
+    intlSpy.calls.reset();
+    spectator.setHostInput({
+      data: 123,
+      config: { useGrouping: false, maximumFractionDigits: 3 }
+    });
+    const second = spectator.element.textContent;
+
+    expect(intlSpy).not.toHaveBeenCalled();
+    expect(second).toBe(first);
+  });
 });