## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: https://github.com/ngneat/transloco/blob/master/CONTRIBUTING.md#commit
- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[x] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or link to a relevant issue. -->

Issue Number: #632

## What is the new behavior?

Locale pipes return last formatted value when value and parameters do not change.


## Does this PR introduce a breaking change?

```
[ ] Yes
[x] No
```

<!-- If this PR contains a breaking change, please describe the impact and migration path for existing applications below. -->

## Other information

Using this measurement code (and same for Date, Percent and Currency):
```typescript
    pipe = new TranslocoDecimalPipe(service, cdr, defaultConfig.localeConfig);

    const config = { useGrouping: true, maximumFractionDigits: 3 };
    
    const start = performance.now();
    for(let i = 0; i < 100000; i++) {
      pipe.transform('123', config);
    }
    const end = performance.now();
    console.log(`Decimal Execution time: ${end - start} ms`);
```

Before optimization I get:
```
LOG: 'Decimal Execution time: 1708.9000000059605 ms'
LOG: 'Date Execution time: 7880.100000008941 ms'
LOG: 'Percent Execution time: 1741.4000000059605 ms'
LOG: 'Currency Execution time: 2233.5 ms'
```

After optimization I get:
```
LOG: 'Decimal Execution time: 35.19999998807907 ms'
LOG: 'Date Execution time: 47.599999994039536 ms'
LOG: 'Percent Execution time: 34.900000005960464 ms'
LOG: 'Currency Execution time: 38.70000000298023 ms'
```